hide directly with checkbox
Also set the checkbuttons checked correctly to begin with (whether user is on the list to hide always). Here a little hack is used since we can't be sure in what order the messages that set up the users on the lists come.

Page reloading is still needed to unhide users that are removed from the list.